+++ src/test/java/com/fillumina/performance/producer/progression/BaseMagnitudeProgressionPerformanceInstrumenterTest.java
@@ -1,6 +1,5 @@
 package com.fillumina.performance.producer.progression;
 
-import com.fillumina.performance.producer.progression.ProgressionPerformanceInstrumenter;
 import com.fillumina.performance.PerformanceTimerBuilder;
 import org.junit.Test;
 
@@ -52,7 +51,7 @@ public class BaseMagnitudeProgressionPerformanceInstrumenterTest {
             return this;
         }
 
-        private void assertSamples(final long... iterations) {
+        private void assertSamples(final int... iterations) {
             final AssertIterationsPerformanceConsumer assertIterations =
                     new AssertIterationsPerformanceConsumer()
                         .setIterations(iterations)
